Rename EventStream and move to its own module.
